Start rewrite of redux slices
