+++ src/redux/deviceDataSlice.ts
@@ -1,0 +1,6 @@
+type DeviceDataSliceType = {
+  deviceId: string;
+  demoModeEnabled: boolean;
+  pushToken: string;
+  getsNotifications: boolean;
+};
