+++ src/redux/marathonSlice.ts
@@ -1,0 +1,5 @@
+import { FirestoreHour } from '../types/FirebaseTypes';
+
+type MarathonSliceType = {
+  marathonHours: FirestoreHour[];
+};
